--- src/modules/chillouts/commands/data/index.ts
@@ -1,5 +1,0 @@
-import addChillCreator from './addChillCreator';
-
-export {
-    addChillCreator,
-};
