+++ src/modules/twitch/commands/data/interactions/addNotification.ts
@@ -2,10 +2,10 @@ import { TwitchGuildPerms } from '@/entities/twitch/twitchGuilds.entity';
 import subscription from '@/entities/twitch/subscription.entity';
 import { ChatInputCommandInteraction, } from 'discord.js';
 import { TModuleContentInfo } from '@/types/client';
-import { TwitchService } from '../../data/services';
+import { TwitchService } from '../../../data/services';
 import { ModuleContentTypes } from '@/constants';
 import { twitchGuild } from '@/entities/twitch';
-import { addNotification, } from '../data';
+import { addNotification, } from '..';
 import modClient from '@/modClient';
 
 export default async function (interaction: ChatInputCommandInteraction, client: modClient) {
