--- src/modules/twitch/commands/data/manager.ts
@@ -1,14 +1,0 @@
-import { CommandType } from "@/modules/commands/data/constants";
-import { PermissionFlagsBits, SlashCommandBuilder, } from "discord.js"
-
-const command = new SlashCommandBuilder()
-    .setName('twitch_manager')
-    .setDescription('...')
-    .setDefaultMemberPermissions(PermissionFlagsBits.ManageChannels);
-
-export default {
-    command,
-    extraInfo: {
-        type: CommandType.GLOBAL,
-    },
-};
