--- src/modules/voiceCreator/commands/data/index.ts
@@ -1,5 +1,0 @@
-import addVoiceCreator from './addVoiceCreator';
-
-export {
-    addVoiceCreator,
-};
